+++ src/main/java/com/tonywww/jeioptimize/integration/SophisticatedStorageShulkerRecipeAccess.java
@@ -1,0 +1,7 @@
+package com.tonywww.jeioptimize.integration;
+
+import net.minecraft.world.item.crafting.ShapelessRecipe;
+
+public interface SophisticatedStorageShulkerRecipeAccess {
+    ShapelessRecipe jeiOptimize$getCompose();
+}